fix(gui): refactor code in UsageDialogPlus
