+++ jadx-gui/src/main/java/jadx/gui/ui/dialog/UsageDialogPlus.java
@@ -5,8 +5,6 @@ import java.awt.Color;
 import java.awt.Component;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
-import java.awt.Point;
-import java.awt.Rectangle;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.util.ArrayList;
@@ -19,7 +17,6 @@ import java.util.Map;
 import javax.swing.BorderFactory;
 import javax.swing.Box;
 import javax.swing.BoxLayout;
-import javax.swing.Icon;
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
 import javax.swing.JLabel;
@@ -33,16 +30,12 @@ import javax.swing.SwingUtilities;
 import javax.swing.WindowConstants;
 import javax.swing.event.TreeExpansionEvent;
 import javax.swing.event.TreeExpansionListener;
-import javax.swing.event.TreeSelectionEvent;
-import javax.swing.event.TreeSelectionListener;
 import javax.swing.tree.DefaultMutableTreeNode;
-import javax.swing.tree.DefaultTreeCellRenderer;
 import javax.swing.tree.DefaultTreeModel;
 import javax.swing.tree.TreeNode;
 import javax.swing.tree.TreePath;
 import javax.swing.tree.TreeSelectionModel;
 
-import org.fife.ui.rsyntaxtextarea.RSyntaxTextArea;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
@@ -63,8 +56,9 @@ import jadx.gui.treemodel.JField;
 import jadx.gui.treemodel.JMethod;
 import jadx.gui.treemodel.JNode;
 import jadx.gui.ui.MainWindow;
-import jadx.gui.ui.codearea.AbstractCodeArea;
+import jadx.gui.ui.cellrenders.PathHighlightTreeCellRenderer;
 import jadx.gui.ui.panel.ProgressPanel;
+import jadx.gui.ui.panel.SimpleCodePanel;
 import jadx.gui.utils.JNodeCache;
 import jadx.gui.utils.NLS;
 import jadx.gui.utils.UiUtils;
@@ -74,14 +68,14 @@ public class UsageDialogPlus extends CommonSearchDialog {
 
 	private final JPanel mainPanel;
 	private final JSplitPane splitPane;
-	private final CodePanel codePanel;
+	private final SimpleCodePanel simpleCodePanel;
 	private final JTree usageTree;
 	private final DefaultTreeModel treeModel;
 	private final DefaultMutableTreeNode rootNode;
 	private final ProgressPanel localProgressPanel;
 	private final JLabel resultsInfoLabel;
 	private final JLabel progressInfoLabel;
-	private JNode initialNode;
+	private final JNode initialNode;
 
 	private static final Logger LOG = LoggerFactory.getLogger(UsageDialogPlus.class);
 
@@ -116,15 +110,15 @@ public class UsageDialogPlus extends CommonSearchDialog {
 		warnLabel.setVisible(false);
 
 		// Initialize result and progress info labels
-		resultsInfoLabel = new JLabel("");
-		progressInfoLabel = new JLabel("");
+		resultsInfoLabel = new JLabel();
+		progressInfoLabel = new JLabel();
 		localProgressPanel = new ProgressPanel(mainWindow, false);
 
 		mainPanel = new JPanel();
 		mainPanel.setLayout(new BorderLayout());
 
 		// Create the code panel
-		codePanel = new CodePanel(mainWindow);
+		simpleCodePanel = new SimpleCodePanel(mainWindow);
 
 		// Create the tree
 		rootNode = new DefaultMutableTreeNode(node);
@@ -136,31 +130,29 @@ public class UsageDialogPlus extends CommonSearchDialog {
 		usageTree.putClientProperty("JTree.lineStyle", "Horizontal");
 		usageTree.setRowHeight(22);
 		usageTree.setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
+		usageTree.setFont(mainWindow.getSettings().getFont());
 
 		// Use a custom renderer instead of a custom UI
 		usageTree.setCellRenderer(new PathHighlightTreeCellRenderer());
 
 		// Add tree listeners
-		usageTree.addTreeSelectionListener(new TreeSelectionListener() {
-			@Override
-			public void valueChanged(TreeSelectionEvent e) {
-				DefaultMutableTreeNode node = (DefaultMutableTreeNode) usageTree.getLastSelectedPathComponent();
-				if (node == null) {
-					return;
-				}
-
-				Object nodeInfo = node.getUserObject();
-				if (nodeInfo instanceof CodeNode) {
-					CodeNode codeNode = (CodeNode) nodeInfo;
-					codePanel.showCode(codeNode, codeNode.makeDescString());
-				} else if (nodeInfo instanceof JNode) {
-					JNode jNode = (JNode) nodeInfo;
-					codePanel.showCode(jNode, jNode.makeDescString());
-				}
-
-				// Update the result information to display the number of child nodes of the currently selected node
-				updateResultsInfo(node);
+		usageTree.addTreeSelectionListener(e -> {
+			DefaultMutableTreeNode node1 = (DefaultMutableTreeNode) usageTree.getLastSelectedPathComponent();
+			if (node1 == null) {
+				return;
 			}
+
+			Object nodeInfo = node1.getUserObject();
+			if (nodeInfo instanceof CodeNode) {
+				CodeNode codeNode = (CodeNode) nodeInfo;
+				simpleCodePanel.showCode(codeNode, codeNode.makeDescString());
+			} else if (nodeInfo instanceof JNode) {
+				JNode jNode = (JNode) nodeInfo;
+				simpleCodePanel.showCode(jNode, jNode.makeDescString());
+			}
+
+			// Update the result information to display the number of child nodes of the currently selected node
+			updateResultsInfo(node1);
 		});
 
 		usageTree.addTreeExpansionListener(new TreeExpansionListener() {
@@ -206,7 +198,7 @@ public class UsageDialogPlus extends CommonSearchDialog {
 				usageTree.setSelectionPath(path);
 
 				// Handle the right-click menu
-				if (e.getButton() == MouseEvent.BUTTON3) {
+				if (SwingUtilities.isRightMouseButton(e)) {
 					DefaultMutableTreeNode selectedNode = (DefaultMutableTreeNode) path.getLastPathComponent();
 					Object userObject = selectedNode.getUserObject();
 
@@ -218,7 +210,7 @@ public class UsageDialogPlus extends CommonSearchDialog {
 				}
 
 				// Handle left-click single/double click
-				if (e.getButton() == MouseEvent.BUTTON1) {
+				if (SwingUtilities.isLeftMouseButton(e)) {
 					long clickTime = System.currentTimeMillis();
 					// Double-click interval
 					long doubleClickInterval = 300;
@@ -299,7 +291,7 @@ public class UsageDialogPlus extends CommonSearchDialog {
 		leftPanel.add(statusPanel, BorderLayout.SOUTH);
 
 		splitPane.setLeftComponent(leftPanel);
-		splitPane.setRightComponent(codePanel);
+		splitPane.setRightComponent(simpleCodePanel);
 
 		mainPanel.add(splitPane, BorderLayout.CENTER);
 
@@ -547,14 +539,10 @@ public class UsageDialogPlus extends CommonSearchDialog {
 		});
 
 		JMenuItem jumpToItem = new JMenuItem(NLS.str("usage_dialog_plus.jump_to"));
-		jumpToItem.addActionListener(evt -> {
-			openItem(node);
-		});
+		jumpToItem.addActionListener(evt -> openItem(node));
 
 		JMenuItem copyPathItem = new JMenuItem(NLS.str("usage_dialog_plus.copy_path"));
-		copyPathItem.addActionListener(evt -> {
-			copyUsagePath(path);
-		});
+		copyPathItem.addActionListener(evt -> copyUsagePath(path));
 
 		popup.add(expandItem);
 		popup.addSeparator();
@@ -594,274 +582,6 @@ public class UsageDialogPlus extends CommonSearchDialog {
 
 			// Copy to clipboard
 			UiUtils.copyToClipboard(pathBuilder.toString());
-		}
-	}
-
-	// Delete the PathHighlightTreeUI class, use the enhanced CellRenderer instead
-	private class PathHighlightTreeCellRenderer extends DefaultTreeCellRenderer {
-		private static final int INDENT_PER_LEVEL = 18; // Each level of indentation in pixels
-		private final Color selectedPathColor = new Color(220, 240, 255); // Light blue background
-		private final Color selectedPathTextColor = new Color(0, 0, 150); // Dark blue text
-
-		@Override
-		public Component getTreeCellRendererComponent(JTree tree, Object value, boolean selected,
-				boolean expanded, boolean leaf, int row, boolean hasFocus) {
-
-			Component comp = super.getTreeCellRendererComponent(tree, value, selected, expanded, leaf, row, hasFocus);
-
-			DefaultMutableTreeNode node = (DefaultMutableTreeNode) value;
-			Object userObject = node.getUserObject();
-
-			// Calculate the node level
-			int level = node.getLevel();
-			// Set different colors according to the level
-			float hue = (level * 0.1f) % 1.0f; // Hue cycle
-			Color levelColor = Color.getHSBColor(hue, 0.1f, 0.95f);
-
-			// Check if it is on the selected path
-			boolean onSelectionPath = false;
-			TreePath selectionPath = tree.getSelectionPath();
-			if (selectionPath != null) {
-				// Check if the current node is on the selected path (whether it is part of the selected path)
-				Object[] selectedPathNodes = selectionPath.getPath();
-				for (Object pathNode : selectedPathNodes) {
-					if (pathNode == node) {
-						onSelectionPath = true;
-						break;
-					}
-				}
-			}
-
-			if (onSelectionPath && !selected) {
-				// If it is on the selected path but not the selected node, use a special background
-				setBackground(selectedPathColor);
-				setForeground(selectedPathTextColor);
-
-				// Set the highlighted border
-				setBorder(BorderFactory.createCompoundBorder(
-						BorderFactory.createLineBorder(selectedPathTextColor, 1),
-						BorderFactory.createEmptyBorder(1, level * 2, 1, 1)));
-			} else if (!selected) {
-				// Only apply the background color when it is not selected
-				setBackground(levelColor);
-				// Normal border
-				setBorder(BorderFactory.createEmptyBorder(2, level * 2 + 1, 2, 1));
-			} else {
-				// The selected node also adds a border
-				setBorder(BorderFactory.createEmptyBorder(2, level * 2 + 1, 2, 1));
-			}
-
-			if (userObject instanceof JNode) {
-				JNode jNode = (JNode) userObject;
-				setText(jNode.makeLongString());
-				setIcon(jNode.getIcon());
-				setToolTipText(jNode.getTooltip());
-			} else if (userObject instanceof CodeNode) {
-				CodeNode codeNode = (CodeNode) userObject;
-				setText(codeNode.makeLongString());
-				setIcon(codeNode.getIcon());
-				setToolTipText(codeNode.getTooltip());
-			}
-
-			return comp;
-		}
-
-		// Add custom icons to enhance visual distinction
-		@Override
-		public Icon getLeafIcon() {
-			return super.getLeafIcon();
-		}
-
-		@Override
-		public Icon getOpenIcon() {
-			return super.getOpenIcon();
-		}
-
-		@Override
-		public Icon getClosedIcon() {
-			return super.getClosedIcon();
-		}
-	}
-
-	// The code panel class is used to display the code of the selected node.
-	private class CodePanel extends JPanel {
-		private final RSyntaxTextArea codeArea;
-		private final JLabel titleLabel;
-
-		public CodePanel(MainWindow mainWindow) {
-			setLayout(new BorderLayout(5, 5));
-			setBorder(BorderFactory.createEmptyBorder(5, 5, 5, 5));
-
-			// Set the minimum size to ensure the panel is not completely minimized
-			setMinimumSize(new Dimension(300, 400));
-			setPreferredSize(new Dimension(800, 600));
-
-			// The title label
-			titleLabel = new JLabel(NLS.str("usage_dialog_plus.code_view"));
-			titleLabel.setFont(codeFont);
-			titleLabel.setBorder(BorderFactory.createEmptyBorder(5, 5, 10, 5));
-
-			// The code area
-			codeArea = AbstractCodeArea.getDefaultArea(mainWindow);
-			codeArea.setEditable(false);
-			codeArea.setText("// " + NLS.str("usage_dialog_plus.select_node"));
-
-			JScrollPane scrollPane = new JScrollPane(codeArea);
-			scrollPane.setHorizontalScrollBarPolicy(JScrollPane.HORIZONTAL_SCROLLBAR_AS_NEEDED);
-			scrollPane.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_AS_NEEDED);
-
-			add(titleLabel, BorderLayout.NORTH);
-			add(scrollPane, BorderLayout.CENTER);
-		}
-
-		public void showCode(JNode node, String codeLine) {
-			if (node != null) {
-				titleLabel.setText(NLS.str("usage_dialog_plus.code_for") + " " + node.makeLongString());
-				codeArea.setSyntaxEditingStyle(node.getSyntaxName());
-
-				// Get the complete code
-				String contextCode = getContextCode(node, codeLine);
-				codeArea.setText(contextCode);
-
-				// Highlight the key line and scroll to that position
-				scrollToCodeLine(codeArea, codeLine);
-
-				// If it is a CodeNode, we can get a more precise position
-				if (node instanceof CodeNode) {
-					CodeNode codeNode = (CodeNode) node;
-					int pos = codeNode.getPos();
-					if (pos > 0) {
-						// Try to use the position information to more accurately locate
-						try {
-							String text = codeArea.getText();
-							int lineNum = 0;
-							int curPos = 0;
-							// Calculate the line number corresponding to the position
-							for (int i = 0; i < text.length() && curPos <= pos; i++) {
-								if (text.charAt(i) == '\n') {
-									lineNum++;
-								}
-								curPos++;
-							}
-
-							if (lineNum > 0) {
-								// Scroll to the calculated line number
-								int finalLineNum = lineNum;
-								SwingUtilities.invokeLater(() -> {
-									try {
-										Rectangle lineRect = codeArea
-												.modelToView(codeArea.getLineStartOffset(finalLineNum));
-										if (lineRect != null) {
-											JScrollPane scrollPane = (JScrollPane) codeArea.getParent().getParent();
-											Rectangle viewRect = scrollPane.getViewport().getViewRect();
-											int y = lineRect.y - (viewRect.height - lineRect.height) / 2;
-											if (y < 0) {
-												y = 0;
-											}
-											scrollPane.getViewport().setViewPosition(new Point(0, y));
-										}
-									} catch (Exception e) {
-										// Fall back to using string matching
-										scrollToCodeLine(codeArea, codeLine);
-									}
-								});
-							}
-						} catch (Exception e) {
-							// Fall back to using string matching
-							scrollToCodeLine(codeArea, codeLine);
-						}
-					} else {
-						// If there is no position information, use string matching
-						scrollToCodeLine(codeArea, codeLine);
-					}
-				} else {
-					// Not a CodeNode, use string matching
-					scrollToCodeLine(codeArea, codeLine);
-				}
-			} else {
-				titleLabel.setText(NLS.str("usage_dialog_plus.code_view"));
-				codeArea.setText("// " + NLS.str("usage_dialog_plus.select_node"));
-			}
-		}
-
-		private String getContextCode(JNode node, String codeLine) {
-			// Always try to get the complete code
-			if (node instanceof CodeNode) {
-				CodeNode codeNode = (CodeNode) node;
-				JNode usageJNode = codeNode.getJParent();
-				if (usageJNode != null) {
-					// Try to get the complete code of the method or class
-					String fullCode = getFullNodeCode(usageJNode);
-					if (fullCode != null && !fullCode.isEmpty()) {
-						return fullCode;
-					}
-				}
-			}
-
-			// If you cannot get more context, at least add some empty lines and comments
-			return "// Unable to get complete context, only display related lines\n\n" + codeLine;
-		}
-
-		private String getFullNodeCode(JNode node) {
-			if (node != null) {
-				// Get the code information of the node
-				ICodeInfo codeInfo = node.getCodeInfo();
-				if (codeInfo != null && !codeInfo.equals(ICodeInfo.EMPTY)) {
-					return codeInfo.getCodeStr();
-				}
-
-				// If it is a class node, try to get the class code
-				if (node instanceof JClass) {
-					JClass jClass = (JClass) node;
-					return jClass.getCodeInfo().getCodeStr();
-				}
-			}
-			return null;
-		}
-
-		private void scrollToCodeLine(RSyntaxTextArea textArea, String lineToHighlight) {
-			// Try to find and highlight a specific line in the code and scroll to that position
-			try {
-				String fullText = textArea.getText();
-				int lineIndex = fullText.indexOf(lineToHighlight);
-				if (lineIndex >= 0) {
-					// Ensure the text area has updated the layout
-					textArea.revalidate();
-
-					// Highlight the code line
-					textArea.setCaretPosition(lineIndex);
-					int endIndex = lineIndex + lineToHighlight.length();
-					textArea.select(lineIndex, endIndex);
-					textArea.getCaret().setSelectionVisible(true);
-
-					// Use SwingUtilities.invokeLater to ensure the scroll is executed after the UI is updated
-					SwingUtilities.invokeLater(() -> {
-						try {
-							// Get the line number
-							int lineNum = textArea.getLineOfOffset(lineIndex);
-							// Ensure the line is centered in the view
-							Rectangle lineRect = textArea.modelToView(textArea.getLineStartOffset(lineNum));
-							if (lineRect != null) {
-								// Calculate the center point of the view
-								JScrollPane scrollPane = (JScrollPane) textArea.getParent().getParent();
-								Rectangle viewRect = scrollPane.getViewport().getViewRect();
-								int y = lineRect.y - (viewRect.height - lineRect.height) / 2;
-								if (y < 0) {
-									y = 0;
-								}
-								// Scroll to the calculated position
-								scrollPane.getViewport().setViewPosition(new Point(0, y));
-							}
-						} catch (Exception e) {
-							LOG.debug("Error scrolling to line: {}", e.getMessage());
-						}
-					});
-				} else {
-					LOG.debug("Could not find line to highlight: {}", lineToHighlight);
-				}
-			} catch (Exception e) {
-				LOG.debug("Error highlighting line: {}", e.getMessage());
-			}
 		}
 	}
