+++ jadx-gui/src/main/resources/i18n/Messages_en_US.properties
@@ -136,7 +136,7 @@ message.unable_preview_font=Unable preview font
 
 heapUsage.text=JADX memory usage: %.2f GB of %.2f GB
 
-common_dialog.ok=Ok
+common_dialog.ok=OK
 common_dialog.cancel=Cancel
 common_dialog.add=Add
 common_dialog.update=Update
@@ -179,7 +179,7 @@ search_dialog.keep_open=Keep open
 search_dialog.tip_searching=Searching
 search_dialog.limit_package=Limit to package:
 search_dialog.package_not_found=No matching package found
-search_dialog.copy=copy all
+search_dialog.copy=Copy All
 
 usage_dialog.title=Usage search
 usage_dialog.label=Usage for:
@@ -188,10 +188,10 @@ usage_dialog_plus.title=Usage tree search
 usage_dialog_plus.jump_to=Jump to current location
 usage_dialog_plus.copy_path=Copy usage tree path
 usage_dialog_plus.search_complete=Search completed
-usage_dialog_plus.code_view=code view
-usage_dialog_plus.select_node=select node
-usage_dialog_plus.code_for=code for
-usage_dialog_plus.expand_usages=expand data
+usage_dialog_plus.code_view=Code View
+usage_dialog_plus.select_node=Select Node
+usage_dialog_plus.code_for=Code for %s
+usage_dialog_plus.expand_usages=Expand Data
 
 comment_dialog.title.add=Add code comment
 comment_dialog.title.update=Update code comment
