+++ jadx-gui/src/main/resources/i18n/Messages_ko_KR.properties
@@ -188,10 +188,10 @@ usage_dialog.label=다음의 사용 검색 결과:
 #usage_dialog_plus.jump_to=Jump to current location
 #usage_dialog_plus.copy_path=Copy usage tree path
 #usage_dialog_plus.search_complete=Search completed
-#usage_dialog_plus.code_view=code view
-#usage_dialog_plus.select_node=select node
-#usage_dialog_plus.code_for=code for
-#usage_dialog_plus.expand_usages=expand data
+#usage_dialog_plus.code_view=Code View
+#usage_dialog_plus.select_node=Select Node
+#usage_dialog_plus.code_for=Code for %s
+#usage_dialog_plus.expand_usages=Expand Data
 
 comment_dialog.title.add=주석 추가
 comment_dialog.title.update=주석 업데이트
