+++ jadx-gui/src/main/resources/i18n/Messages_pt_BR.properties
@@ -188,10 +188,10 @@ usage_dialog.label=Usado por:
 #usage_dialog_plus.jump_to=Jump to current location
 #usage_dialog_plus.copy_path=Copy usage tree path
 #usage_dialog_plus.search_complete=Search completed
-#usage_dialog_plus.code_view=code view
-#usage_dialog_plus.select_node=select node
-#usage_dialog_plus.code_for=code for
-#usage_dialog_plus.expand_usages=expand data
+#usage_dialog_plus.code_view=Code View
+#usage_dialog_plus.select_node=Select Node
+#usage_dialog_plus.code_for=Code for %s
+#usage_dialog_plus.expand_usages=Expand Data
 
 comment_dialog.title.add=Adicionar comentário ao código
 comment_dialog.title.update=Atualizar comentário do código
