+++ jadx-gui/src/main/resources/i18n/Messages_zh_CN.properties
@@ -190,7 +190,7 @@ usage_dialog_plus.copy_path=复制用例路径
 usage_dialog_plus.search_complete=搜索完成
 usage_dialog_plus.code_view=代码预览
 usage_dialog_plus.select_node=选择节点
-usage_dialog_plus.code_for=代码信息
+usage_dialog_plus.code_for=代码信息 %s
 usage_dialog_plus.expand_usages=展开数据
 
 comment_dialog.title.add=添加代码注释
