+++ jadx-gui/src/main/resources/i18n/Messages_zh_TW.properties
@@ -190,7 +190,7 @@ usage_dialog_plus.copy_path=複製使用情況路徑
 usage_dialog_plus.search_complete=搜尋完成
 usage_dialog_plus.code_view=程式碼預覽
 usage_dialog_plus.select_node=選擇節點
-usage_dialog_plus.code_for=程式碼資訊
+usage_dialog_plus.code_for=程式碼資訊 %s
 usage_dialog_plus.expand_usages=展開資料
 
 comment_dialog.title.add=新增程式碼註解
